% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.r
\name{bench}
\alias{bench}
\alias{benchR6}
\title{bench class}
\usage{
bench(env = parent.frame())
}
\arguments{
\item{env}{Environment to execute expressions in.}
}
\value{
A bench class object.
}
\description{
Storage and methods benchmarking data.

Constructor for bench objects.
}
\examples{
library(merkhet)

b = bench()

b$time(Sys.sleep(.2), 2)
b
b$time(Sys.sleep(.35))
b

b$csv()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{benchR6$new()}}
\item \href{#method-time}{\code{benchR6$time()}}
\item \href{#method-print}{\code{benchR6$print()}}
\item \href{#method-csv}{\code{benchR6$csv()}}
\item \href{#method-plot}{\code{benchR6$plot()}}
\item \href{#method-clone}{\code{benchR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{benchR6$new(env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{Environment to execute expressions in.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Class initializer.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-time"></a>}}
\if{latex}{\out{\hypertarget{method-time}{}}}
\subsection{Method \code{time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{benchR6$time(expr, reps = 1, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{Expression to time.}

\item{\code{reps}}{Number of replications for each expression.}

\item{\code{name}}{Optional name for the timed expression.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Time an expression using \code{system.time()}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{benchR6$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print current benchmark information.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-csv"></a>}}
\if{latex}{\out{\hypertarget{method-csv}{}}}
\subsection{Method \code{csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{benchR6$csv(sep = ",", header = TRUE, file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{Separator.}

\item{\code{header}}{Print header?}

\item{\code{file}}{A file to write to, or \code{NULL} to print to stdout.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Prints the benchmark table as a csv.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{benchR6$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to \code{barplot()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Show a histogram of the benchmarks table.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{benchR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
