% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merkeht-package.R
\docType{package}
\name{merkhet-package}
\alias{merkhet-package}
\title{Merkhet:  Time-Related Utilities for R}
\description{
The merkhet package contains several time-related utilities for
benchmarkers and those interested in performance.
}
\details{
The name 'merkhet' is an ancient word for a kind of time-keeping 
device, which means the 'instrument of knowing.'

\tabular{ll}{ 
   Package: \tab merkhet \cr 
   Type: \tab Package \cr 
   License: \tab BSD 2-Clause \cr 
   NeedsCompilation: \tab yes 
}
}
\references{
Project home page: \url{http://github.com/shinra-dev/merkhet}
}
\author{
Drew Schmidt
}
\keyword{Package}
