% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readabletime.r
\name{print-readabletime}
\alias{print-readabletime}
\alias{print.readabletime}
\title{print.readabletime}
\usage{
\method{print}{readabletime}(x, ..., digits = 3)
}
\arguments{
\item{x}{An readabletime object.}

\item{...}{Additional arguments (ignored).}

\item{digits}{The (max) number of digits to display.}
}
\description{
Print readabletiem objects.
}
