% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.r
\name{print-sysinfo}
\alias{print-sysinfo}
\alias{print.sysinfo}
\title{print.sysinfo}
\usage{
\method{print}{sysinfo}(x, ...)
}
\arguments{
\item{x}{An sysinfo object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for sysinfo objects.
}

