% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.r
\name{systime}
\alias{systime}
\alias{Sys.systime}
\title{systime}
\usage{
Sys.systime()
}
\value{
A readabletime object.
}
\description{
Returns the amount of time in human-readable units that the current process
has been in kernel mode.
}
\examples{
\dontrun{
library(memuse)

Sys.systime()
}

}
