% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.r
\name{uptime}
\alias{uptime}
\alias{Sys.uptime}
\title{uptime}
\usage{
Sys.uptime()
}
\value{
A readabletime object.
}
\description{
Returns the amount of time to the nearest second in human-readable units
since the system was booted.
}
\examples{
\dontrun{
library(memuse)

Sys.uptime()
}

}
