% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.r
\name{usrtime}
\alias{usrtime}
\alias{Sys.usrtime}
\alias{Sys.runtime}
\title{usrtime}
\usage{
Sys.usrtime()

Sys.runtime()
}
\value{
A readabletime object.

A readabletime object.
}
\description{
Returns the amount of time in human-readable units that the current process
has been in user mode.

Returns the amount of time to the nearest second in human-readable units that
the current R process has been active, i.e. the wall-clock time of the
process.
}
\examples{
\dontrun{
library(memuse)

Sys.usrtime()
}

\dontrun{
library(memuse)

Sys.runtime()
}

}
