% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merkeht-package.R
\docType{package}
\name{merkhet-package}
\alias{merkhet-package}
\title{Merkhet:  Time Utilities for R}
\description{
The name 'merkhet' is an ancient word for a kind of timekeeping
device, which means the "instrument of knowing." The package contains
several timing utilities useful for 'benchmarking' and performance
monitoring.
}
\details{
\tabular{ll}{ 
   Package: \tab merkhet \cr 
   Type: \tab Package \cr 
   License: \tab BSD 2-Clause \cr 
   NeedsCompilation: \tab yes 
}
}
\references{
Project home page: \url{http://github.com/shinra-dev/merkhet}
}
\author{
Drew Schmidt
}
\keyword{Package}
