% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merkeht-package.R
\docType{package}
\name{merkhet-package}
\alias{merkhet-package}
\title{Merkhet:  Time-Related Utilities for R}
\description{
TODO
}
\details{
\tabular{ll}{
   Package: \tab merkhet \cr
   Type: \tab Package \cr
   License: \tab BSD 2-Clause \cr
   NeedsCompilation: \tab yes
}
}
\author{
Drew Schmidt
}
\references{
Project home page: \url{http://github.com/shinra-dev/merkhet}
}
\keyword{Package}

