% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.r
\name{post-system-time}
\alias{post-system-time}
\alias{post.system.time}
\title{post.system.time}
\usage{
post.system.time()
}
\value{
A proc_time object (same as \code{system.time()}).
}
\description{
Returns the user, system (kernel), and wallclock time that the current
process has been active. Like a kind of post-hoc \code{system.time()}.
}
\details{
If any of the user, system, or elapsed process times are unsupported on a
given platform, then the value will be \code{NA}.
}
\examples{
\dontrun{
library(memuse)

post.system.time()
}

}
